initial

+++ VictronModbus/module.php
@@ -123,7 +123,7 @@ class VictronModbus extends Module
                 elseif ($config['scale'] == 0) {
                     $value = $value[1].$value[2].$value[3].$value[4].$value[5].$value[6];
                 } elseif ($config['scale'] == 1) {
-                    $value = (float)$value[1];
+                    $value = (int)$value[1];
                 } elseif ($config['scale'] == 10) {
                     $value = (float)$value[1]/10;
                 } elseif ($config['scale'] == 100) {
